modularized theming for scaling
