+++ app/src/components/Registration/Container/styles.ts
@@ -1,0 +1,7 @@
+export const styles = (theme: any) => ({
+    //example of additional classes for conditional styles used with theming
+    //this may not be a great example as the baseline colors are defaulting to our theming, it is simply to illustrate
+    container: {
+        color: theme.palette.primary.main,
+    },
+})
